requirements.txt
# 네이버 블로그 크롤러 필요 패키지
playwright==1.58.0
beautifulsoup4==4.14.3
asyncio==4.0.0
lxml==6.0.2
requests==2.32.5

# 데이터 처리
pandas==2.3.3
numpy==2.4.2

# 기타 유틸리티
python-dateutil==2.9.0.post0
pytz==2025.2

# 추천 API (recommendation/*.py, server GET /recommend)
python-dotenv>=1.0.0
supabase>=2.3.0
openai>=1.40.0
httpx>=0.27.0
